Many updates

+++ PathFinderTest/Tests/Interactive/IWorldWriter.cs
@@ -4,7 +4,7 @@ namespace PathFinderTest.Tests.Interactive
 {
     public interface IWorldWriter
     {
-        void WriteResult(int testNum, decimal thoroughness, double cost, int ticks, long cpuCycles);
+        void WriteResult(int testNum, double thoroughness, double cost, int ticks, long cpuCycles);
         void WriteInfo(string info);
         void DrawPosition(int x, int y);
         void DrawPosition(int x, int y, int testNumber);
